requirements.txt
boto3==1.16.57
python-dotenv==0.15.0
